@@ -11,19 +11,18 @@ android {
         }
     }
 
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.taisheng.now"
         minSdkVersion 19
-        targetSdkVersion 22
+        targetSdkVersion 28
         versionCode 9
         versionName "1.6"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
-        jackOptions {
-            enabled true
-        }
+//        jackOptions {
+//            enabled true
+//        }
 
         ndk {
             abiFilters "armeabi","armeabi-v7a","arm64-v8a"
@@ -43,6 +42,9 @@ android {
     }
 
 
+
+
+
     compileOptions {
         sourceCompatibility 1.8
         targetCompatibility 1.8
@@ -50,30 +52,32 @@ android {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:22.2.1'
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:multidex:1.0.1'
-    compile 'com.tencent.bugly:crashreport_upgrade:1.3.4'
-    compile 'org.greenrobot:eventbus:3.0.0'
-    compile 'com.squareup.retrofit2:retrofit:2.1.0'
-    compile 'com.squareup.okhttp3:logging-interceptor:3.4.1'
-    compile 'org.ligboy.retrofit2:converter-fastjson:2.1.0'
-    compile 'com.facebook.fresco:fresco:0.12.0'
-    compile 'com.android.support:design:23.3.0'
-    compile 'com.tencent.liteav:LiteAVSDK_TRTC:latest.release'
-    compile 'com.android.support:support-v4:26.0.0-alpha1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+//    testCompile 'junit:junit:4.12'
+    implementation 'androidx.multidex:multidex:2.0.0'
+    implementation 'com.tencent.bugly:crashreport_upgrade:1.3.4'
+    implementation 'org.greenrobot:eventbus:3.0.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.1.0'
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.4.1'
+    implementation 'org.ligboy.retrofit2:converter-fastjson:2.1.0'
+    implementation 'com.facebook.fresco:fresco:0.12.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'com.tencent.liteav:LiteAVSDK_TRTC:latest.release'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     //小米推送
-    compile files('libs/MiPush_SDK_Client_3_6_18.jar')
+    implementation files('libs/MiPush_SDK_Client_3_6_18.jar')
     //首先引入gson库
-    compile 'com.google.code.gson:gson:2.8.1'
-    compile 'com.android.support:recyclerview-v7:24.2.1'
-    compile 'com.zzhoujay.richtext:richtext:2.4.7'
+    implementation 'com.google.code.gson:gson:2.8.1'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'com.zzhoujay.richtext:richtext:2.4.7'
 
 
     //下拉刷新
-    compile 'in.srain.cube:ultra-ptr:1.0.11'
+    implementation 'in.srain.cube:ultra-ptr:1.0.11'
+
+    implementation project(':libskywebrtc')
 }
