@@ -1,9 +1,10 @@
 package com.taisheng.now.base;
 
 import android.app.Activity;
-import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+
+import androidx.fragment.app.Fragment;
+//import android.support.v4.app.Fragment;
 
 
 /**
