@@ -1,9 +1,11 @@
 package com.taisheng.now.view.gallery;
 
 import android.content.Context;
-import android.support.v4.view.PagerAdapter;
+//import android.support.v4.view.PagerAdapter;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.viewpager.widget.PagerAdapter;
 
 import com.taisheng.now.R;
 
